# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: ploykit API
  version: 0.3.0
  description: |
    ploykit 框架 + example 产品应用的 HTTP 契约。

    认证两通道（webx.Authenticate）：
    - 浏览器会话 Cookie `tk_auth`（写请求须带 X-CSRF-Token，/config 拉取 masked token）；
    - PAT `Authorization: Bearer tk_*`（豁免 CSRF）。

    扁平路由（/api/billing/*、/api/usage、/api/audit*、/api/tasks*）的工作区上下文
    由 `X-Workspace-Id` 请求头注入（@ploykit/client 自动携带）。

    错误信封统一为 ErrorBody（webx.ErrorBody）：`{"error":"E_*","message":string,"details"?}`。
security:
  - sessionCookie: []
  - patBearer: []
...
paths:
...
  /auth/send-code:
    post:
      operationId: sendCode
      tags: [auth]
      summary: 发送登录验证码（dev 模式回 DEV_CODE）
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendCodeRequest'
      responses:
        '200':
          description: 已发送（无论邮箱是否存在均回同值，防枚举）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusSent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimited'
        default:
          $ref: '#/components/responses/DefaultError'
  /auth/verify-code:
    post:
      operationId: verifyCode
      tags: [auth]
      summary: 验证码登录（成功签发会话 Cookie tk_auth）
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyCodeRequest'
      responses:
        '200':
          description: 登录成功（Set-Cookie tk_auth）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '429':
          $ref: '#/components/responses/RateLimited'
        default:
          $ref: '#/components/responses/DefaultError'
...
  /api/workspaces/{workspaceId}/sso:
    put:
      operationId: putWorkspaceSSO
      tags: [workspace]
      summary: 配置工作区 OIDC 联邦（owner/admin）
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SSOConfigRequest'
      responses:
        '200':
          description: 已保存的联邦配置（secret 脱敏）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SSOConfig'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          $ref: '#/components/responses/Unavailable'
        default:
          $ref: '#/components/responses/DefaultError'
    get:
      operationId: getWorkspaceSSO
      tags: [workspace]
      summary: 查看工作区 OIDC 联邦配置（owner/admin；secret 脱敏）
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      responses:
        '200':
          description: 联邦配置
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SSOConfig'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/DefaultError'
    delete:
      operationId: deleteWorkspaceSSO
      tags: [workspace]
      summary: 移除工作区 OIDC 联邦配置（owner/admin）
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      responses:
        '200':
          description: 已移除
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusOK'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/DefaultError'
...
  /auth/me:
    get:
      operationId: getMe
      tags: [session]
      summary: 当前用户
      description: |
        模拟会话（管理员 impersonation 签发）额外透出 impersonated_by
        （发起管理员 userID，ADR 0008）；普通会话无此字段。
      responses:
        '200':
          description: 当前用户资料
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PKUser'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        default:
          $ref: '#/components/responses/DefaultError'
    patch:
      operationId: updateMe
      tags: [session]
      summary: 更新资料（昵称/头像）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMeRequest'
      responses:
        '200':
          description: 更新后的用户
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PKUser'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        default:
          $ref: '#/components/responses/DefaultError'
    delete:
      operationId: deleteMe
      tags: [session]
      summary: 账号注销（软删除，仅浏览器会话）
      description: |
        软删除当前账号：display_name 置空、密码/PAT 全吊销、tokens_valid_after
        双闸使当前会话即刻失效，响应同时清会话 Cookie。幂等二调（用户已注销）
        返回 404。PAT 调用返回 403（仅浏览器会话可执行）。
      security:
        - sessionCookie: []
      responses:
        '204':
          description: 已注销
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/DefaultError'
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [session]
      summary: 修改密码（成功后轮换会话 Cookie，其余会话吊销；仅浏览器会话）
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: 修改成功（Set-Cookie 新会话）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpiresAtResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        default:
          $ref: '#/components/responses/DefaultError'
  /auth/sessions:
    get:
      operationId: listSessions
      tags: [session]
      summary: 当前用户活跃会话列表（仅浏览器会话）
      security:
        - sessionCookie: []
      responses:
        '200':
          description: 会话列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SessionInfo'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        default:
          $ref: '#/components/responses/DefaultError'
    delete:
      operationId: revokeAllSessions
      tags: [session]
      summary: 吊销全部会话（含当前，登出所有设备；仅浏览器会话）
      security:
        - sessionCookie: []
      responses:
        '200':
          description: 已全部吊销
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusOK'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        default:
          $ref: '#/components/responses/DefaultError'
  /auth/sessions/{sessionId}:
    delete:
      operationId: revokeSession
      tags: [session]
      summary: 吊销单个会话（设备管理；仅浏览器会话）
      description: |
        吊销本人名下指定会话。吊销"当前会话"等价登出（后端不特判、不清
        Cookie，前端按 GET /auth/sessions 的 current 标记提示处理）。
        不存在/已吊销/非本人 → 404（重复吊销幂等）。PAT 调用返回 403
        （仅浏览器会话可执行）。
      security:
        - sessionCookie: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema: { type: string }
          description: 会话 ID（GET /auth/sessions 返回的 id）
      responses:
        '204':
          description: 已吊销
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/DefaultError'

  # ---------- identity：PAT ----------
  /api/tokens:
    get:
      operationId: listPATs
      tags: [pat]
      summary: PAT 列表（仅浏览器会话）
      security:
        - sessionCookie: []
      responses:
        '200':
          description: PAT 列表（不含密钥本体）
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PAT'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        default:
          $ref: '#/components/responses/DefaultError'
    post:
      operationId: createPAT
      tags: [pat]
      summary: 创建 PAT（token 仅本次响应出现一次；仅浏览器会话）
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePATRequest'
      responses:
        '201':
          description: 已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatePATResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        default:
          $ref: '#/components/responses/DefaultError'
  /api/tokens/{id}:
    delete:
      operationId: revokePAT
      tags: [pat]
      summary: 吊销 PAT（仅浏览器会话）
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/TokenId'
      responses:
        '200':
          description: 已吊销
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusOK'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/DefaultError'

  # ---------- workspace：我的工作区 ----------
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: tk_auth
      description: 浏览器会话 Cookie（webx.Authenticate 签发/续期；写请求须另带 X-CSRF-Token）
    patBearer:
      type: http
      scheme: bearer
      description: '个人访问令牌（Authorization: Bearer tk_*，CSRF 豁免）'
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema: { type: string }
      description: 工作区 ID
...
    TokenId:
      name: id
      in: path
      required: true
      schema: { type: string }
      description: PAT ID
...
  responses:
    DefaultError:
      description: 其他错误（统一 ErrorBody 信封）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    BadRequest:
      description: 参数非法（E_VALIDATION / E_BAD_JSON / E_PAYLOAD_TOO_LARGE）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    Unauthenticated:
      description: 未认证（E_UNAUTHENTICATED）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    Forbidden:
      description: 无权限（E_FORBIDDEN）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    NotFound:
      description: 不存在（E_NOT_FOUND）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
...
    RateLimited:
      description: 限流（E_RATE_LIMITED）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
...
    Unavailable:
      description: 服务未接线/不可用（E_UNAVAILABLE）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
  schemas:
...
    ErrorBody:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: 错误码（E_UNAUTHENTICATED / E_FORBIDDEN / E_NOT_FOUND / E_VALIDATION / E_CONFLICT / E_RATE_LIMITED / E_QUOTA_EXCEEDED / E_INTERNAL / E_TIMEOUT / E_PAYLOAD_TOO_LARGE / E_UNAVAILABLE）
        message:
          type: string
        details:
          description: 可选补充（通常为字符串详情）
    StatusOK:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ok]
    StatusSent:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [sent]
...
    PKUser:
      type: object
      required: [id, email, display_name, status, is_platform_admin, email_verified, created_at]
      properties:
        id:
          type: string
        email:
          type: string
        display_name:
          type: string
        avatar_url:
          type: string
          description: 空串或 http(s):// 前缀且 ≤2048 字节（与前端 zod 同语义）
        status:
          type: string
          enum: [active, disabled, deleted]
        is_platform_admin:
          type: boolean
        email_verified:
          type: boolean
        created_at:
          type: string
        impersonated_by:
          type: string
          description: |
            仅 GET /auth/me 出现且仅在模拟会话下（发起模拟的平台管理员
            userID；session.impersonated_by，迁移 034 / ADR 0008）。普通
            会话省略该字段。登录/注册响应不含此字段。
          format: date-time
    LoginResponse:
      type: object
      required: [user, expires_at]
      properties:
        user:
          $ref: '#/components/schemas/PKUser'
        expires_at:
          type: string
          format: date-time
    ExpiresAtResponse:
      type: object
      required: [expires_at]
      properties:
        expires_at:
          type: string
          format: date-time
    SessionInfo:
      type: object
      required: [id, ip_hash, user_agent, last_seen_at, expires_at, created_at]
      properties:
        id:
          type: string
        ip_hash:
          type: string
        user_agent:
          type: string
        last_seen_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    PAT:
      type: object
      required: [id, name, prefix, created_at]
      properties:
        id:
          type: string
        name:
          type: string
        prefix:
          type: string
          description: 展示用前缀（密钥本体不回传）
        expires_at:
          type: string
          format: date-time
        last_used_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    CreatePATResponse:
      type: object
      required: [pat, token]
      properties:
        pat:
          $ref: '#/components/schemas/PAT'
        token:
          type: string
          description: 完整令牌（tk_*，仅本次出现一次）
    SendCodeRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
    VerifyCodeRequest:
      type: object
      required: [email, code]
      properties:
        email:
          type: string
        code:
          type: string
...
    UpdateMeRequest:
      type: object
      required: [display_name, avatar_url]
      properties:
        display_name:
          type: string
          minLength: 1
          maxLength: 100
          description: trim 后 1..100 字符（rune 计）
        avatar_url:
          type: string
          maxLength: 2048
          description: 空串（清除头像）或 http(s):// 前缀 URL，≤2048 字节（与前端 zod 同语义）
    ChangePasswordRequest:
      type: object
      required: [old_password, new_password]
      properties:
        old_password:
          type: string
        new_password:
          type: string
    CreatePATRequest:
      type: object
      required: [name, ttl_hours]
      properties:
        name:
          type: string
        ttl_hours:
          type: integer
          description: 有效期小时数（0 = 不过期）

    # ---- workspace（PKWorkspace 对齐 packages/ui） ----
...
    SSOConfigRequest:
      type: object
      required: [issuer_url, client_id, client_secret]
      properties:
        issuer_url:
          type: string
          description: IdP 基址（自动 discovery .well-known/openid-configuration）
        client_id:
          type: string
        client_secret:
          type: string
        scopes:
          type: string
          default: openid email profile
    SSOConfig:
      type: object
      required: [workspace_id, issuer_url, client_id, client_secret, scopes, updated_at]
      properties:
        workspace_id:
          type: string
        issuer_url:
          type: string
        client_id:
          type: string
        client_secret:
          type: string
          description: 恒为 ***（脱敏）
        scopes:
          type: string
        updated_at:
          type: string
          format: date-time
...